src/requirements.txt
fastapi==0.115.6
SQLAlchemy==2.0.24
pydantic-settings==2.3.4
uvicorn
aiosqlite
python-jose[cryptography]
passlib[bcrypt]